back end added, front end 50%

+++ book_club_organizer/src/main/java/com/pial/book_club_organizer/service/BookService.java
@@ -1,12 +1,13 @@
 package com.pial.book_club_organizer.service;
 
+import com.pial.book_club_organizer.exception.BookAlreadyExistsException;
+import com.pial.book_club_organizer.exception.BookNotFoundException;
 import com.pial.book_club_organizer.model.Book;
 import com.pial.book_club_organizer.repository.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class BookService {
@@ -17,21 +18,33 @@ public class BookService {
         return bookRepository.findAll();
     }
 
-    public Optional<Book> getBookById(Integer id) {
-        return bookRepository.findById(id);
+    public Book getBookById(Integer id) throws BookNotFoundException {
+        return bookRepository.findById(id).orElseThrow(()->new BookNotFoundException("Book not found!"));
     }
 
-    public void addBook(Book book) {
+    public void addBook(Book book) throws BookAlreadyExistsException {
+        if (bookRepository.existsByTitle(book.getTitle())) {
+            throw new BookAlreadyExistsException("Book already exists!");
+        }
         bookRepository.save(book);
     }
 
-    public void updateBook(Book book) {
+    public void updateBook(Book book) throws BookNotFoundException {
+        if (!bookRepository.existsById(book.getId())) {
+            throw new BookNotFoundException("Book not found!");
+        }
         bookRepository.saveAndFlush(book);
     }
 
 
-    public void deleteBookById(Integer id) {
+    public void deleteBookById(Integer id) throws BookNotFoundException {
+        if (!bookRepository.existsById(id)) {
+            throw new BookNotFoundException("Book not found!");
+        }
         bookRepository.deleteById(id);
     }
 
+    public List<Book> findByTitleContainingOrAuthorContainingOrGenreContaining(String query, String query1, String query2) {
+        return bookRepository.findByTitleContainingOrAuthorContainingOrGenreContaining(query,query1,query2);
+    }
 }
